In order to enable proper testing and speed up the development process (particularly when adding end user customisation options) the code base is being refactored to use a large main package and more flexible parsing of user configuration options. This should allow for much faster development speed in the future.
This is the correct time to focus now that the software has proven to work and has seen some (rather more than I ever expected) public use.

In addition to the refactor, this commit includes improvements to the logging process and some expanded user customisations in the form of: selecting an existing pbf file, selecting an out dir separate from the dataDir, choices on the inclusion of ocean and landuse data and a max ram count.

In due course, improvements to the splitting algorithms are planned which will greatly increase processing speed and allow for finer control on the amount of ram to be used (down to extremely low amounts for the task at hand) while also allowing the amount of ram to be used to be increased in line with available system resources so that no time is wasted when generating tiles.